Move partitioning stuff to separate dir tiling
