+++ roms_tools/setup/grid.py
@@ -8,6 +8,7 @@ import matplotlib.pyplot as plt
 import yaml
 import importlib.metadata
 from typing import Dict, Union, List
+from roms_tools.utils import save_datasets
 from roms_tools.setup.topography import _add_topography
 from roms_tools.setup.mask import _add_mask, _add_velocity_masks
 from roms_tools.vertical_coordinate import (
@@ -22,7 +23,7 @@ from roms_tools.setup.utils import (
     get_target_coords,
     gc_dist,
 )
-from roms_tools.setup.utils import extract_single_value, save_datasets
+from roms_tools.setup.utils import extract_single_value
 from pathlib import Path
