+++ roms_tools/setup/utils.py
@@ -3,7 +3,6 @@ import numpy as np
 from typing import Union, Any, Dict, Type
 import pandas as pd
 import cftime
-import logging
 from pathlib import Path
 from datetime import datetime
 from dataclasses import fields, asdict
@@ -580,42 +579,6 @@ def group_by_year(ds, filepath):
         output_filenames.append(output_filename)
 
     return dataset_list, output_filenames
-
-
-def save_datasets(dataset_list, output_filenames, use_dask=False):
-    """Save the list of datasets to netCDF4 files.
-
-    Parameters
-    ----------
-    dataset_list : list
-        List of datasets to be saved.
-    output_filenames : list
-        List of filenames for the output files.
-    use_dask : bool, optional
-        Whether to use Dask diagnostics (e.g., progress bars) when saving the datasets, by default False.
-
-    Returns
-    -------
-    List[Path]
-        A list of Path objects for the filenames that were saved.
-    """
-
-    saved_filenames = []
-
-    output_filenames = [f"{filename}.nc" for filename in output_filenames]
-    logging.info("Saving dataset to NetCDF files: %s", ", ".join(output_filenames))
-
-    if use_dask:
-        from dask.diagnostics import ProgressBar
-
-        with ProgressBar():
-            xr.save_mfdataset(dataset_list, output_filenames)
-    else:
-        xr.save_mfdataset(dataset_list, output_filenames)
-
-    saved_filenames.extend(Path(f) for f in output_filenames)
-
-    return saved_filenames
 
 
 def get_target_coords(grid, use_coarse_grid=False):
