+++ roms_tools/tests/test_tiling/test_partition.py
@@ -2,7 +2,7 @@ import pytest
 from pathlib import Path
 import xarray.testing as xrt
 
-from roms_tools.utils import partition, partition_netcdf
+from roms_tools.tiling.partition import partition, partition_netcdf
 from roms_tools import Grid
